Redis_added
