+++ .gitignore
@@ -163,3 +163,7 @@ logs/BotLogger
 /tg_bot/logs/BotLogger/
 /postgres/
 /admin_panel/media/
+
+#vscode
+.vscode
+/vscode
