+++ requirements.txt
@@ -2,3 +2,5 @@ aiogram==3.4.1
 environs==11.0.0
 redis==5.0.0
 psycopg2-binary==2.9.9
+apscheduler==3.10.4
+apscheduler-di==0.1.0
